@@ -8,16 +8,15 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "barryvdh/laravel-translation-manager": "^0.5.7",
+        "php": "^7.3",
         "doctrine/dbal": "^2.10",
         "fideloper/proxy": "^4.0",
         "laracasts/flash": "^3.0",
         "laracasts/presenter": "^0.2.2",
-        "laravel/framework": "^7.0",
-        "laravel/socialite": "^4.4",
+        "laravel/framework": "8.x-dev",
+        "laravel/socialite": "^5.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
+        "laravel/ui": "^3.0",
         "laravelcollective/html": "^6.0",
         "snapappointments/bootstrap-select": "^1.13",
         "spatie/laravel-activitylog": "^3.14",
@@ -26,15 +25,15 @@
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.6",
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.6",
         "fruitcake/laravel-telescope-toolbar": "^1.1",
         "fzaninotto/faker": "^1.4",
-        "laravel/telescope": "^3.0",
+        "laravel/telescope": "^4.0",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^4.1",
+        "nunomaduro/collision": "^5.0",
         "orangehill/iseed": "^2.6",
-        "phpunit/phpunit": "^8.5",
-        "tanmuhittin/laravel-google-translate": "^1.0"
+        "phpunit/phpunit": "^9.0",
+        "tanmuhittin/laravel-google-translate": "^2.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -48,12 +47,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
