@@ -22,10 +22,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -48,6 +52,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -109,6 +116,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -116,6 +125,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -124,6 +134,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -132,11 +143,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -177,7 +183,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -199,10 +204,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -225,6 +234,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -286,6 +298,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -293,6 +307,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -301,6 +316,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -309,11 +325,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -354,7 +365,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -376,10 +386,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -402,6 +416,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -463,6 +480,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -470,6 +489,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -478,6 +498,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -486,11 +507,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -531,7 +547,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -553,10 +568,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -579,6 +598,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -640,6 +662,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -647,6 +671,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -655,6 +680,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -663,11 +689,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -708,7 +729,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -730,10 +750,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -756,6 +780,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -817,6 +844,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -824,6 +853,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -832,6 +862,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -840,11 +871,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -885,7 +911,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -907,10 +932,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -933,6 +962,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -994,6 +1026,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -1001,6 +1035,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -1009,6 +1044,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -1017,11 +1053,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -1062,7 +1093,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -1084,10 +1114,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -1110,6 +1144,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -1171,6 +1208,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -1178,6 +1217,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -1186,6 +1226,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -1194,11 +1235,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -1239,7 +1275,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -1261,10 +1296,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -1287,6 +1326,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -1348,6 +1390,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -1355,6 +1399,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -1363,6 +1408,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -1371,11 +1417,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -1416,7 +1457,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -1438,10 +1478,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -1464,6 +1508,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -1525,6 +1572,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -1532,6 +1581,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -1540,6 +1590,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -1548,11 +1599,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -1593,7 +1639,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -1615,10 +1660,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -1641,6 +1690,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -1702,6 +1754,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -1709,6 +1763,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -1717,6 +1772,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -1725,11 +1781,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -1770,7 +1821,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
@@ -1792,10 +1842,14 @@ namespace PHPSTORM_META {
             'Fruitcake\TelescopeToolbar\Toolbar' => \Fruitcake\TelescopeToolbar\Toolbar::class,
             'Illuminate\Auth\Middleware\RequirePassword' => \Illuminate\Auth\Middleware\RequirePassword::class,
             'Illuminate\Broadcasting\BroadcastManager' => \Illuminate\Broadcasting\BroadcastManager::class,
+            'Illuminate\Bus\BatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
+            'Illuminate\Bus\DatabaseBatchRepository' => \Illuminate\Bus\DatabaseBatchRepository::class,
             'Illuminate\Bus\Dispatcher' => \Illuminate\Bus\Dispatcher::class,
+            'Illuminate\Cache\RateLimiter' => \Illuminate\Cache\RateLimiter::class,
             'Illuminate\Console\Scheduling\Schedule' => \Illuminate\Console\Scheduling\Schedule::class,
             'Illuminate\Console\Scheduling\ScheduleFinishCommand' => \Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
             'Illuminate\Console\Scheduling\ScheduleRunCommand' => \Illuminate\Console\Scheduling\ScheduleRunCommand::class,
+            'Illuminate\Console\Scheduling\ScheduleWorkCommand' => \Illuminate\Console\Scheduling\ScheduleWorkCommand::class,
             'Illuminate\Contracts\Auth\Access\Gate' => \Illuminate\Auth\Access\Gate::class,
             'Illuminate\Contracts\Broadcasting\Broadcaster' => \Illuminate\Broadcasting\Broadcasters\LogBroadcaster::class,
             'Illuminate\Contracts\Console\Kernel' => \App\Console\Kernel::class,
@@ -1818,6 +1872,9 @@ namespace PHPSTORM_META {
             'NunoMaduro\Collision\Contracts\Provider' => \NunoMaduro\Collision\Provider::class,
             'Spatie\Activitylog\ActivityLogStatus' => \Spatie\Activitylog\ActivityLogStatus::class,
             'Spatie\Activitylog\ActivityLogger' => \Spatie\Activitylog\ActivityLogger::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Contracts\ApiTranslatorContract' => \Tanmuhittin\LaravelGoogleTranslate\Api\StichozaApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslate::class,
+            'Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute' => \Tanmuhittin\LaravelGoogleTranslate\Translators\ApiTranslateWithAttribute::class,
             'Whoops\Handler\HandlerInterface' => \Facade\Ignition\ErrorPage\IgnitionWhoopsHandler::class,
             'auth' => \Illuminate\Auth\AuthManager::class,
             'auth.driver' => \Illuminate\Auth\SessionGuard::class,
@@ -1879,6 +1936,8 @@ namespace PHPSTORM_META {
             'command.package.discover' => \Illuminate\Foundation\Console\PackageDiscoverCommand::class,
             'command.policy.make' => \Illuminate\Foundation\Console\PolicyMakeCommand::class,
             'command.provider.make' => \Illuminate\Foundation\Console\ProviderMakeCommand::class,
+            'command.queue.batches-table' => \Illuminate\Queue\Console\BatchesTableCommand::class,
+            'command.queue.clear' => \Illuminate\Queue\Console\ClearCommand::class,
             'command.queue.failed' => \Illuminate\Queue\Console\ListFailedCommand::class,
             'command.queue.failed-table' => \Illuminate\Queue\Console\FailedTableCommand::class,
             'command.queue.flush' => \Illuminate\Queue\Console\FlushFailedCommand::class,
@@ -1886,6 +1945,7 @@ namespace PHPSTORM_META {
             'command.queue.listen' => \Illuminate\Queue\Console\ListenCommand::class,
             'command.queue.restart' => \Illuminate\Queue\Console\RestartCommand::class,
             'command.queue.retry' => \Illuminate\Queue\Console\RetryCommand::class,
+            'command.queue.retry-batch' => \Illuminate\Queue\Console\RetryBatchCommand::class,
             'command.queue.table' => \Illuminate\Queue\Console\TableCommand::class,
             'command.queue.work' => \Illuminate\Queue\Console\WorkCommand::class,
             'command.request.make' => \Illuminate\Foundation\Console\RequestMakeCommand::class,
@@ -1894,6 +1954,7 @@ namespace PHPSTORM_META {
             'command.route.clear' => \Illuminate\Foundation\Console\RouteClearCommand::class,
             'command.route.list' => \Illuminate\Foundation\Console\RouteListCommand::class,
             'command.rule.make' => \Illuminate\Foundation\Console\RuleMakeCommand::class,
+            'command.schema.dump' => \Illuminate\Database\Console\DumpCommand::class,
             'command.seed' => \Illuminate\Database\Console\Seeds\SeedCommand::class,
             'command.seeder.make' => \Illuminate\Database\Console\Seeds\SeederMakeCommand::class,
             'command.serve' => \Illuminate\Foundation\Console\ServeCommand::class,
@@ -1902,11 +1963,6 @@ namespace PHPSTORM_META {
             'command.stub.publish' => \Illuminate\Foundation\Console\StubPublishCommand::class,
             'command.test.make' => \Illuminate\Foundation\Console\TestMakeCommand::class,
             'command.tinker' => \Laravel\Tinker\Console\TinkerCommand::class,
-            'command.translation-manager.clean' => \Barryvdh\TranslationManager\Console\CleanCommand::class,
-            'command.translation-manager.export' => \Barryvdh\TranslationManager\Console\ExportCommand::class,
-            'command.translation-manager.find' => \Barryvdh\TranslationManager\Console\FindCommand::class,
-            'command.translation-manager.import' => \Barryvdh\TranslationManager\Console\ImportCommand::class,
-            'command.translation-manager.reset' => \Barryvdh\TranslationManager\Console\ResetCommand::class,
             'command.up' => \Illuminate\Foundation\Console\UpCommand::class,
             'command.vendor.publish' => \Illuminate\Foundation\Console\VendorPublishCommand::class,
             'command.view.cache' => \Illuminate\Foundation\Console\ViewCacheCommand::class,
@@ -1947,7 +2003,6 @@ namespace PHPSTORM_META {
             'router' => \Illuminate\Routing\Router::class,
             'session' => \Illuminate\Session\SessionManager::class,
             'session.store' => \Illuminate\Session\Store::class,
-            'translation-manager' => \Barryvdh\TranslationManager\Manager::class,
             'translation.loader' => \Illuminate\Translation\FileLoader::class,
             'translator' => \Illuminate\Translation\Translator::class,
             'url' => \Illuminate\Routing\UrlGenerator::class,
