@@ -1,31 +1,32 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Models\Page;
-use Faker\Generator as Faker;
-use Illuminate\Support\Str;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+class PageFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Page::class;
 
-$factory->define(Page::class, function (Faker $faker, $data) {
-    return [
-        'title'         => trans('model.title'),
-        'content'       => '<p>' . trans('model.content') . '</p>',
-        'is_first'      => $data['is_first'] ?? false,
-        'is_last'       => false,
-        'is_checkpoint' => false,
-        'story_id'      => $data['story_id'],
-        'layout'        => 'play1',
-    ];
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'title'         => trans('model.title'),
+            'content'       => '<p>' . trans('model.content') . '</p>',
+            'is_last'       => false,
+            'is_checkpoint' => false,
+            'layout'        => 'play1',
+        ];
+    }
 }
-);
