@@ -1,36 +1,30 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
-use App\Models\Story;
-use Faker\Generator as Faker;
-use Illuminate\Support\Str;
+use App\Models\Page;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+class StoryFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Page::class;
 
-$factory->define(Story::class, function (Faker $faker, $data) {
-
-    $return = [
-        'title'        => ucfirst($faker->words(5, true)),
-        'description'  => $faker->text,
-        'user_id'      => $data['user_id'],
-        'is_published' => true,
-    ];
-
-    return [
-        'title'        => ucfirst($faker->words(5, true)),
-        'description'  => $faker->text,
-        'user_id'      => $data['user_id'],
-        'is_published' => true,
-    ];
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'title'        => ucfirst($this->faker->words(5, true)),
+            'description'  => $this->faker->text,
+            'is_published' => true,
+        ];
+    }
 }
-);
