@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class ActivityLogTableSeeder extends Seeder
@@ -12,12 +14,12 @@ class ActivityLogTableSeeder extends Seeder
      */
     public function run()
     {
-        
+
 
         \DB::table('activity_log')->delete();
-        
+
         \DB::table('activity_log')->insert(array (
-            0 => 
+            0 =>
             array (
                 'id' => 0,
                 'log_name' => 'new_game',
@@ -30,7 +32,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 14:09:59',
                 'updated_at' => '2020-08-14 14:09:59',
             ),
-            1 => 
+            1 =>
             array (
                 'id' => 45,
                 'log_name' => 'new_game',
@@ -43,7 +45,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 04:49:59',
                 'updated_at' => '2020-08-07 04:49:59',
             ),
-            2 => 
+            2 =>
             array (
                 'id' => 46,
                 'log_name' => 'reset_game',
@@ -56,7 +58,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 04:50:23',
                 'updated_at' => '2020-08-07 04:50:23',
             ),
-            3 => 
+            3 =>
             array (
                 'id' => 47,
                 'log_name' => 'new_game',
@@ -69,7 +71,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:06',
                 'updated_at' => '2020-08-07 05:12:06',
             ),
-            4 => 
+            4 =>
             array (
                 'id' => 48,
                 'log_name' => 'end_game',
@@ -82,7 +84,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:08',
                 'updated_at' => '2020-08-07 05:12:08',
             ),
-            5 => 
+            5 =>
             array (
                 'id' => 49,
                 'log_name' => 'reset_game',
@@ -95,7 +97,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:12',
                 'updated_at' => '2020-08-07 05:12:12',
             ),
-            6 => 
+            6 =>
             array (
                 'id' => 50,
                 'log_name' => 'new_game',
@@ -108,7 +110,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:15',
                 'updated_at' => '2020-08-07 05:12:15',
             ),
-            7 => 
+            7 =>
             array (
                 'id' => 51,
                 'log_name' => 'end_game',
@@ -121,7 +123,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:19',
                 'updated_at' => '2020-08-07 05:12:19',
             ),
-            8 => 
+            8 =>
             array (
                 'id' => 52,
                 'log_name' => 'new_game',
@@ -134,7 +136,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:23',
                 'updated_at' => '2020-08-07 05:12:23',
             ),
-            9 => 
+            9 =>
             array (
                 'id' => 53,
                 'log_name' => 'end_game',
@@ -147,7 +149,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:25',
                 'updated_at' => '2020-08-07 05:12:25',
             ),
-            10 => 
+            10 =>
             array (
                 'id' => 54,
                 'log_name' => 'reset_game',
@@ -160,7 +162,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-07 05:12:28',
                 'updated_at' => '2020-08-07 05:12:28',
             ),
-            11 => 
+            11 =>
             array (
                 'id' => 57,
                 'log_name' => 'new_game',
@@ -173,7 +175,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 13:02:34',
                 'updated_at' => '2020-08-11 13:02:34',
             ),
-            12 => 
+            12 =>
             array (
                 'id' => 58,
                 'log_name' => 'end_game',
@@ -186,7 +188,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 13:02:47',
                 'updated_at' => '2020-08-11 13:02:47',
             ),
-            13 => 
+            13 =>
             array (
                 'id' => 59,
                 'log_name' => 'reset_game',
@@ -199,7 +201,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 13:02:50',
                 'updated_at' => '2020-08-11 13:02:50',
             ),
-            14 => 
+            14 =>
             array (
                 'id' => 60,
                 'log_name' => 'new_game',
@@ -212,7 +214,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:08:33',
                 'updated_at' => '2020-08-11 14:08:33',
             ),
-            15 => 
+            15 =>
             array (
                 'id' => 61,
                 'log_name' => 'end_game',
@@ -225,7 +227,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:08:52',
                 'updated_at' => '2020-08-11 14:08:52',
             ),
-            16 => 
+            16 =>
             array (
                 'id' => 62,
                 'log_name' => 'reset_game',
@@ -238,7 +240,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:09:04',
                 'updated_at' => '2020-08-11 14:09:04',
             ),
-            17 => 
+            17 =>
             array (
                 'id' => 63,
                 'log_name' => 'new_game',
@@ -251,7 +253,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:09:06',
                 'updated_at' => '2020-08-11 14:09:06',
             ),
-            18 => 
+            18 =>
             array (
                 'id' => 64,
                 'log_name' => 'end_game',
@@ -264,7 +266,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:09:09',
                 'updated_at' => '2020-08-11 14:09:09',
             ),
-            19 => 
+            19 =>
             array (
                 'id' => 65,
                 'log_name' => 'reset_game',
@@ -277,7 +279,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:09:16',
                 'updated_at' => '2020-08-11 14:09:16',
             ),
-            20 => 
+            20 =>
             array (
                 'id' => 66,
                 'log_name' => 'new_game',
@@ -290,7 +292,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:09:18',
                 'updated_at' => '2020-08-11 14:09:18',
             ),
-            21 => 
+            21 =>
             array (
                 'id' => 67,
                 'log_name' => 'end_game',
@@ -303,7 +305,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:09:19',
                 'updated_at' => '2020-08-11 14:09:19',
             ),
-            22 => 
+            22 =>
             array (
                 'id' => 68,
                 'log_name' => 'reset_game',
@@ -316,7 +318,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-11 14:09:27',
                 'updated_at' => '2020-08-11 14:09:27',
             ),
-            23 => 
+            23 =>
             array (
                 'id' => 69,
                 'log_name' => 'new_game',
@@ -329,7 +331,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:02',
                 'updated_at' => '2020-08-12 13:53:02',
             ),
-            24 => 
+            24 =>
             array (
                 'id' => 70,
                 'log_name' => 'end_game',
@@ -342,7 +344,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:10',
                 'updated_at' => '2020-08-12 13:53:10',
             ),
-            25 => 
+            25 =>
             array (
                 'id' => 71,
                 'log_name' => 'reset_game',
@@ -355,7 +357,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:15',
                 'updated_at' => '2020-08-12 13:53:15',
             ),
-            26 => 
+            26 =>
             array (
                 'id' => 72,
                 'log_name' => 'new_game',
@@ -368,7 +370,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:18',
                 'updated_at' => '2020-08-12 13:53:18',
             ),
-            27 => 
+            27 =>
             array (
                 'id' => 73,
                 'log_name' => 'end_game',
@@ -381,7 +383,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:28',
                 'updated_at' => '2020-08-12 13:53:28',
             ),
-            28 => 
+            28 =>
             array (
                 'id' => 74,
                 'log_name' => 'reset_game',
@@ -394,7 +396,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:35',
                 'updated_at' => '2020-08-12 13:53:35',
             ),
-            29 => 
+            29 =>
             array (
                 'id' => 75,
                 'log_name' => 'new_game',
@@ -407,7 +409,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:37',
                 'updated_at' => '2020-08-12 13:53:37',
             ),
-            30 => 
+            30 =>
             array (
                 'id' => 76,
                 'log_name' => 'end_game',
@@ -420,7 +422,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:39',
                 'updated_at' => '2020-08-12 13:53:39',
             ),
-            31 => 
+            31 =>
             array (
                 'id' => 77,
                 'log_name' => 'reset_game',
@@ -433,7 +435,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:44',
                 'updated_at' => '2020-08-12 13:53:44',
             ),
-            32 => 
+            32 =>
             array (
                 'id' => 78,
                 'log_name' => 'new_game',
@@ -446,7 +448,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:46',
                 'updated_at' => '2020-08-12 13:53:46',
             ),
-            33 => 
+            33 =>
             array (
                 'id' => 79,
                 'log_name' => 'end_game',
@@ -459,7 +461,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 13:53:52',
                 'updated_at' => '2020-08-12 13:53:52',
             ),
-            34 => 
+            34 =>
             array (
                 'id' => 80,
                 'log_name' => 'new_game',
@@ -472,7 +474,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:24:18',
                 'updated_at' => '2020-08-12 15:24:18',
             ),
-            35 => 
+            35 =>
             array (
                 'id' => 81,
                 'log_name' => 'end_game',
@@ -485,7 +487,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:01',
                 'updated_at' => '2020-08-12 15:25:01',
             ),
-            36 => 
+            36 =>
             array (
                 'id' => 82,
                 'log_name' => 'reset_game',
@@ -498,7 +500,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:12',
                 'updated_at' => '2020-08-12 15:25:12',
             ),
-            37 => 
+            37 =>
             array (
                 'id' => 83,
                 'log_name' => 'new_game',
@@ -511,7 +513,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:15',
                 'updated_at' => '2020-08-12 15:25:15',
             ),
-            38 => 
+            38 =>
             array (
                 'id' => 84,
                 'log_name' => 'end_game',
@@ -524,7 +526,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:17',
                 'updated_at' => '2020-08-12 15:25:17',
             ),
-            39 => 
+            39 =>
             array (
                 'id' => 85,
                 'log_name' => 'reset_game',
@@ -537,7 +539,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:22',
                 'updated_at' => '2020-08-12 15:25:22',
             ),
-            40 => 
+            40 =>
             array (
                 'id' => 86,
                 'log_name' => 'new_game',
@@ -550,7 +552,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:25',
                 'updated_at' => '2020-08-12 15:25:25',
             ),
-            41 => 
+            41 =>
             array (
                 'id' => 87,
                 'log_name' => 'end_game',
@@ -563,7 +565,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:27',
                 'updated_at' => '2020-08-12 15:25:27',
             ),
-            42 => 
+            42 =>
             array (
                 'id' => 88,
                 'log_name' => 'reset_game',
@@ -576,7 +578,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:28',
                 'updated_at' => '2020-08-12 15:25:28',
             ),
-            43 => 
+            43 =>
             array (
                 'id' => 89,
                 'log_name' => 'new_game',
@@ -589,7 +591,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:30',
                 'updated_at' => '2020-08-12 15:25:30',
             ),
-            44 => 
+            44 =>
             array (
                 'id' => 90,
                 'log_name' => 'end_game',
@@ -602,7 +604,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:32',
                 'updated_at' => '2020-08-12 15:25:32',
             ),
-            45 => 
+            45 =>
             array (
                 'id' => 91,
                 'log_name' => 'reset_game',
@@ -615,7 +617,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:25:51',
                 'updated_at' => '2020-08-12 15:25:51',
             ),
-            46 => 
+            46 =>
             array (
                 'id' => 92,
                 'log_name' => 'new_game',
@@ -628,7 +630,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:01',
                 'updated_at' => '2020-08-12 15:26:01',
             ),
-            47 => 
+            47 =>
             array (
                 'id' => 93,
                 'log_name' => 'end_game',
@@ -641,7 +643,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:05',
                 'updated_at' => '2020-08-12 15:26:05',
             ),
-            48 => 
+            48 =>
             array (
                 'id' => 94,
                 'log_name' => 'reset_game',
@@ -654,7 +656,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:06',
                 'updated_at' => '2020-08-12 15:26:06',
             ),
-            49 => 
+            49 =>
             array (
                 'id' => 95,
                 'log_name' => 'new_game',
@@ -667,7 +669,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:08',
                 'updated_at' => '2020-08-12 15:26:08',
             ),
-            50 => 
+            50 =>
             array (
                 'id' => 96,
                 'log_name' => 'end_game',
@@ -680,7 +682,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:10',
                 'updated_at' => '2020-08-12 15:26:10',
             ),
-            51 => 
+            51 =>
             array (
                 'id' => 97,
                 'log_name' => 'reset_game',
@@ -693,7 +695,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:12',
                 'updated_at' => '2020-08-12 15:26:12',
             ),
-            52 => 
+            52 =>
             array (
                 'id' => 98,
                 'log_name' => 'new_game',
@@ -706,7 +708,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:13',
                 'updated_at' => '2020-08-12 15:26:13',
             ),
-            53 => 
+            53 =>
             array (
                 'id' => 99,
                 'log_name' => 'end_game',
@@ -719,7 +721,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:15',
                 'updated_at' => '2020-08-12 15:26:15',
             ),
-            54 => 
+            54 =>
             array (
                 'id' => 100,
                 'log_name' => 'reset_game',
@@ -732,7 +734,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-12 15:26:24',
                 'updated_at' => '2020-08-12 15:26:24',
             ),
-            55 => 
+            55 =>
             array (
                 'id' => 155,
                 'log_name' => 'new_game',
@@ -745,7 +747,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 07:26:20',
                 'updated_at' => '2020-08-14 07:26:20',
             ),
-            56 => 
+            56 =>
             array (
                 'id' => 156,
                 'log_name' => 'end_game',
@@ -758,7 +760,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 07:27:13',
                 'updated_at' => '2020-08-14 07:27:13',
             ),
-            57 => 
+            57 =>
             array (
                 'id' => 157,
                 'log_name' => 'reset_game',
@@ -771,7 +773,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 07:27:23',
                 'updated_at' => '2020-08-14 07:27:23',
             ),
-            58 => 
+            58 =>
             array (
                 'id' => 162,
                 'log_name' => 'dead_end',
@@ -784,7 +786,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 14:09:42',
                 'updated_at' => '2020-08-14 14:09:42',
             ),
-            59 => 
+            59 =>
             array (
                 'id' => 163,
                 'log_name' => 'dead_end',
@@ -797,7 +799,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 14:09:49',
                 'updated_at' => '2020-08-14 14:09:49',
             ),
-            60 => 
+            60 =>
             array (
                 'id' => 164,
                 'log_name' => 'dead_end',
@@ -810,7 +812,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 14:09:50',
                 'updated_at' => '2020-08-14 14:09:50',
             ),
-            61 => 
+            61 =>
             array (
                 'id' => 172,
                 'log_name' => 'end_game',
@@ -823,7 +825,7 @@ class ActivityLogTableSeeder extends Seeder
                 'created_at' => '2020-08-14 14:09:59',
                 'updated_at' => '2020-08-14 14:09:59',
             ),
-            62 => 
+            62 =>
             array (
                 'id' => 173,
                 'log_name' => 'dead_end',
@@ -837,7 +839,7 @@ class ActivityLogTableSeeder extends Seeder
                 'updated_at' => '2020-08-16 08:16:14',
             ),
         ));
-        
-        
+
+
     }
 }
