@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class CharacterFieldTableSeeder extends Seeder
@@ -12,37 +14,37 @@ class CharacterFieldTableSeeder extends Seeder
      */
     public function run()
     {
-        
+
 
         \DB::table('character_field')->delete();
-        
+
         \DB::table('character_field')->insert(array (
-            0 => 
+            0 =>
             array (
                 'character_id' => 84,
                 'field_id' => 11,
                 'value' => 40,
             ),
-            1 => 
+            1 =>
             array (
                 'character_id' => 84,
                 'field_id' => 12,
                 'value' => 15,
             ),
-            2 => 
+            2 =>
             array (
                 'character_id' => 85,
                 'field_id' => 11,
                 'value' => 6,
             ),
-            3 => 
+            3 =>
             array (
                 'character_id' => 85,
                 'field_id' => 12,
                 'value' => 6,
             ),
         ));
-        
-        
+
+
     }
 }
