@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 
 class DescriptionsTableSeeder extends Seeder
@@ -12,12 +14,12 @@ class DescriptionsTableSeeder extends Seeder
      */
     public function run()
     {
-        
+
 
         \DB::table('descriptions')->delete();
-        
+
         \DB::table('descriptions')->insert(array (
-            0 => 
+            0 =>
             array (
                 'id' => 1,
                 'page_id' => 1,
@@ -25,7 +27,7 @@ class DescriptionsTableSeeder extends Seeder
                 'description' => '<p>Le futur, c\'est quand le passé n\'est plus que de l\'histoire... Ca veut rien dire mais au moins j\'ai un texte long et qui pourra tout faire bugguer si je me plante :p</p>',
                 'deleted_at' => NULL,
             ),
-            1 => 
+            1 =>
             array (
                 'id' => 2,
                 'page_id' => 1,
@@ -33,7 +35,7 @@ class DescriptionsTableSeeder extends Seeder
                 'description' => '<p>Les innovations c\'est compliqué...</p><p>Mais bon, on s\'en fout après tout :D</p>',
                 'deleted_at' => NULL,
             ),
-            2 => 
+            2 =>
             array (
                 'id' => 3,
                 'page_id' => 74,
@@ -41,7 +43,7 @@ class DescriptionsTableSeeder extends Seeder
                 'description' => '<p>Aucune poignée, pas de gonds que tu puisses démonter, tu as beau pousser rien ne se passe.</p>',
                 'deleted_at' => NULL,
             ),
-            3 => 
+            3 =>
             array (
                 'id' => 4,
                 'page_id' => 74,
@@ -49,7 +51,7 @@ class DescriptionsTableSeeder extends Seeder
                 'description' => '<p>Un pavé numérique standard, il faut taper un code à <b>6 chiffres</b> pour débloquer je ne sais quoi.</p>',
                 'deleted_at' => NULL,
             ),
-            4 => 
+            4 =>
             array (
                 'id' => 5,
                 'page_id' => 74,
@@ -57,7 +59,7 @@ class DescriptionsTableSeeder extends Seeder
                 'description' => '<p>La première page est noircie de formules mathématiques compliquées.</p><p>La seconde en revanche contient une liste de nombres premiers : 2, 3, 5, 7, 11 puis le reste jusqu\'à 97. Passionnant.</p>',
                 'deleted_at' => NULL,
             ),
-            5 => 
+            5 =>
             array (
                 'id' => 7,
                 'page_id' => 74,
@@ -65,7 +67,7 @@ class DescriptionsTableSeeder extends Seeder
                 'description' => '<p>Vraiment hideux. L\'auteur doit pourtant en être très fier vu qu\'il l\'a signé : &quot;Renbrante&quot;. Ok, même ça il aurait dû s\'abstenir.</p>',
                 'deleted_at' => NULL,
             ),
-            6 => 
+            6 =>
             array (
                 'id' => 8,
                 'page_id' => 74,
@@ -73,7 +75,7 @@ class DescriptionsTableSeeder extends Seeder
                 'description' => '<p>Un seul vrai livre relié, le reste sont des magazines Picsou, des Picsou Géant. Une bonne cinquantaine au bas mot.</p>',
                 'deleted_at' => NULL,
             ),
-            7 => 
+            7 =>
             array (
                 'id' => 9,
                 'page_id' => 168,
@@ -81,7 +83,7 @@ class DescriptionsTableSeeder extends Seeder
             'description' => '<p>Tu peux écrire pas mal de choses là-dedans, et même mettre en <u>forme </u><i>certaines </i><b style="color: rgb(206, 0, 0);">choses</b>.</p><p>Pratique non ?</p><p style="text-align: right; "><span style="font-style: italic; color: rgb(206, 198, 206);">Le tutorialisateur fou</span></p>',
                 'deleted_at' => NULL,
             ),
-            8 => 
+            8 =>
             array (
                 'id' => 11,
                 'page_id' => 168,
@@ -90,7 +92,7 @@ class DescriptionsTableSeeder extends Seeder
                 'deleted_at' => NULL,
             ),
         ));
-        
-        
+
+
     }
 }
